clean up context loss recovery
